+++ src/constants/instrumentation/common.py
@@ -1,0 +1,18 @@
+TIKTOKEN_MODEL_MAPPING = {
+    "gpt-4": "cl100k_base",
+    "gpt-4-32k": "cl100k_base",
+    "gpt-4-0125-preview": "cl100k_base",
+    "gpt-4-1106-preview": "cl100k_base",
+    "gpt-4-1106-vision-preview": "cl100k_base",
+}
+
+SERVICE_PROVIDERS = {
+    "OPENAI": "OpenAI",
+    "AZURE": "Azure",
+    "LANGCHAIN": "Langchain",
+    "LANGCHAIN_CORE": "Langchain Core",
+    "LANGCHAIN_COMMUNITY": "Langchain Community",
+    "PINECONE": "Pinecone",
+    "LLAMAINDEX": "LlamaIndex",
+    "CHROMA": "Chroma",
+}
