+++ src/entrypoint.py
@@ -1,16 +1,21 @@
 # from examples.langchain_example.basic import basic, load_and_split, rag
 
-# from examples.langchain_example.tool import tool_example
 
+
+# from examples.openai import images_generate
 # from examples.openai.function_calling import function_calling
 
-# from examples.openai.chat_completion import chat_completion
+from examples.openai.chat_completion import chat_completion
+from examples.pinecone_example.basic import basic
 
-from examples.llamaindex_example.basic import basic
 
 # load_and_split()
 # rag()
 basic()
 # chat_completion()
 # function_calling()
-# tool_example()
+# images_generate()
+# embeddings_create()
+
+
+
