+++ src/examples/chroma_example/__init__.py
@@ -1,0 +1,1 @@
+from .basic import basic
