+++ src/examples/chroma_example/basic.py
@@ -2,12 +2,11 @@ import chromadb
 from chromadb.utils import embedding_functions
 from dotenv import find_dotenv, load_dotenv
 
-from examples.setup import setup_instrumentation
-from instrumentation.with_root_span import with_langtrace_root_span
-
+from utils.with_root_span import with_langtrace_root_span
+from init import init
 _ = load_dotenv(find_dotenv())
 
-setup_instrumentation()
+init()
 
 
 @with_langtrace_root_span()
