+++ src/examples/langchain_example/tool.py
@@ -1,17 +1,17 @@
 from dotenv import find_dotenv, load_dotenv
+from init import init
 from langchain import hub
 from langchain.agents import AgentExecutor, create_openai_functions_agent
 from langchain.chains import LLMMathChain
 from langchain_core.pydantic_v1 import BaseModel, Field
 from langchain_core.tools import Tool
 from langchain_openai import ChatOpenAI
+from utils.with_root_span import with_langtrace_root_span
 
-from examples.setup import setup_instrumentation
-from instrumentation.with_root_span import with_langtrace_root_span
 
 _ = load_dotenv(find_dotenv())
 
-setup_instrumentation()
+init()
 
 
 llm = ChatOpenAI(temperature=0, model="gpt-4")
