+++ src/examples/llamaindex_example/basic.py
@@ -1,12 +1,13 @@
 from dotenv import find_dotenv, load_dotenv
+from init import init
 from llama_index.core import SimpleDirectoryReader, VectorStoreIndex
+from utils.with_root_span import with_langtrace_root_span
 
-from examples.setup import setup_instrumentation
-from instrumentation.with_root_span import with_langtrace_root_span
 
 _ = load_dotenv(find_dotenv())
 
-setup_instrumentation()
+
+init()
 
 
 @with_langtrace_root_span()
