+++ src/examples/openai/chat_completion.py
@@ -1,12 +1,13 @@
 from dotenv import find_dotenv, load_dotenv
 from openai import OpenAI
 
-from examples.setup import setup_instrumentation
-from instrumentation.with_root_span import with_langtrace_root_span
+from init.init import init
+from utils.with_root_span import with_langtrace_root_span
+
 
 _ = load_dotenv(find_dotenv())
 
-setup_instrumentation()
+init()
 
 client = OpenAI()
