+++ src/init/__init__.py
@@ -1,0 +1,1 @@
+from init.init import init
