--- src/instrumentation/constants.py
@@ -1,13 +1,0 @@
-"""
-This file contains the constants used in the project.
-"""
-SERVICE_PROVIDERS = {
-    "OPENAI": "OpenAI",
-    "AZURE": "Azure",
-    "LANGCHAIN": "Langchain",
-    "LANGCHAIN_CORE": "Langchain Core",
-    "LANGCHAIN_COMMUNITY": "Langchain Community",
-    "PINECONE": "Pinecone",
-    "LLAMAINDEX": "LlamaIndex",
-    "CHROMA": "Chroma",
-}
