+++ src/instrumentation/llamaindex/patch.py
@@ -5,7 +5,8 @@ from langtrace.trace_attributes import FrameworkSpanAttributes
 from opentelemetry.trace import SpanKind
 from opentelemetry.trace.status import Status, StatusCode
 
-from instrumentation.constants import SERVICE_PROVIDERS
+from constants.instrumentation.common import SERVICE_PROVIDERS
+
 
 
 def generic_patch(method, task, tracer, version):
